Obsidian and other chaincode use (#14)
* Adding Obsidian use.
* Adding Obsidian use without chaincode lifecycle TLS.
* Adding artifact content chaincode.
* Testing an HTTP call from within chaincode.
* Trying the Go IPFS API
* Updating REST API to include blockchain interaction methods.
* Adding artifact graph processing.
* Updated graph population and retrieval.
* Enabled kubernetes dashboard and metrics scraping. Added relationship API methods.
* Updating artifact creation.
* Added API method to get all artifacts.
* Added initial API-based file retrieval.
* Updating to write relationship data via chaincode.
* Create initial scenario for creating related artifacts.
* Added tooling for generating relationship graphs.

+++ hyperledger/scripts/create_chaincode.sh
@@ -5,17 +5,30 @@ mkdir -p $CHAINCODE_TMP_DIR
 
 function package_chaincode_for() {
   local org=$1
-  local cc_folder="../../chaincode/${CHAINCODE_NAME}/ledger"
-  local build_folder="build/chaincode"
-  local cc_archive="${build_folder}/${CHAINCODE_NAME}.tgz"
-  echo "Packaging chaincode folder ${cc_folder}"
+  local ccname=$2
 
-  mkdir -p ${build_folder}
+  local cc_folder="../../chaincode/${ccname}/ledger"
 
-  tar -C ${cc_folder} -zcf ${cc_folder}/code.tar.gz connection.json
-  tar -C ${cc_folder} -zcf ${cc_archive} code.tar.gz metadata.json
+  if [ -d "$cc_folder" ]; then
+    local build_folder="build/chaincode"
+    local cc_archive="${build_folder}/${ccname}.tgz"
+    echo "Packaging chaincode folder ${cc_folder}"
 
-  rm ${cc_folder}/code.tar.gz
+    mkdir -p ${build_folder}
+
+    cat ${cc_folder}/connection.json
+
+    tar -C ${cc_folder} -zcf ${cc_folder}/code.tar.gz connection.json
+
+    cat ${cc_folder}/metadata.json
+
+    tar -C ${cc_folder} -zcf ${cc_archive} code.tar.gz metadata.json
+
+    rm ${cc_folder}/code.tar.gz
+  else
+    echo "Error: ${cc_folder} not found. Can not continue."
+    exit 1
+  fi
 }
 
 # Copy the chaincode archive from the local host to the org admin
@@ -29,7 +42,8 @@ function transfer_chaincode_archive_for() {
   # popd
 
   local org=$1
-  local cc_archive="build/chaincode/${CHAINCODE_NAME}.tgz"
+  local ccname=$2
+  local cc_archive="build/chaincode/${ccname}.tgz"
   echo "Transferring chaincode archive to ${org}"
 
   # Like kubectl cp, but targeted to a deployment rather than an individual pod.
@@ -39,12 +53,13 @@ function transfer_chaincode_archive_for() {
 function install_chaincode_for() {
   local org=$1
   local peer=$2
-  echo "Installing chaincode for ${org} ${peer}"
+  local ccname=$3
+  echo "Installing chaincode ${ccname} for ${org} ${peer}"
 
   # Install the chaincode
   echo 'set -x
   export CORE_PEER_ADDRESS='${org}'-'${peer}':7051
-  peer lifecycle chaincode install build/chaincode/'${CHAINCODE_NAME}'.tgz
+  peer lifecycle chaincode install build/chaincode/'${ccname}'.tgz
   ' | exec kubectl -n $NS exec deploy/${org}-admin-cli -c main -i -- /bin/bash
 
   # local org=$1
@@ -68,51 +83,60 @@ function install_chaincode_for() {
 # Package and install the chaincode, but do not activate.
 function install_chaincode() {
   local org=org1
+  local ccname=$1
 
-  package_chaincode_for ${org}
-  transfer_chaincode_archive_for ${org}
-  install_chaincode_for ${org} peer1
-  install_chaincode_for ${org} peer2
+  package_chaincode_for ${org} ${ccname}
+  transfer_chaincode_archive_for ${org} ${ccname}
+  install_chaincode_for ${org} peer1 ${ccname}
+  install_chaincode_for ${org} peer2 ${ccname}
 
-  set_chaincode_id
+  set_chaincode_id ${ccname}
 }
 
 # Normally the chaincode ID is emitted by the peer install command.  In this case, we'll generate the
 # package ID as the sha-256 checksum of the chaincode archive.
 function set_chaincode_id() {
-  local cc_package=build/chaincode/${CHAINCODE_NAME}.tgz
+  local ccname=$1
+  local cc_package=build/chaincode/${ccname}.tgz
   cc_sha256=$(shasum -a 256 ${cc_package} | tr -s ' ' | cut -d ' ' -f 1)
 
-  label=$( jq -r '.label' ../../chaincode/${CHAINCODE_NAME}/ledger/metadata.json)
+  label=$( jq -r '.label' ../../chaincode/${ccname}/ledger/metadata.json)
 
   export CHAINCODE_ID=${label}:${cc_sha256}
 }
 
 function launch_chaincode_service() {
   local org=$1
-  echo "Launching chaincode container \"${CHAINCODE_IMAGE}\""
+  export PEER_NAME=$2
+  local ccimage=$3
+  local ccname=$4
+  echo "Launching chaincode container \"${ccimage}\""
 
   # The chaincode endpoint needs to have the generated chaincode ID available in the environment.
   # This could be from a config map, a secret, or by directly editing the deployment spec.  Here we'll keep
   # things simple by using sed to substitute script variables into a yaml template.
   export ORG_NUMBER=
-  export PEER_NAME=$2
-  applyPopulatedTemplate ../config/${org}-cc-template.yaml $CHAINCODE_TMP_DIR/${org}-cc.yaml $NS
-  kubectl -n $NS rollout status deploy/${org}-${PEER_NAME}-cc-${CHAINCODE_NAME}
+  export CHAINCODE_IMAGE=$ccimage
+  export CHAINCODE_NAME=$ccname
+
+  applyPopulatedTemplate ../config/${org}-cc-template.yaml $CHAINCODE_TMP_DIR/${org}-${PEER_NAME}-cc-${ccname}.yaml $NS
+  kubectl -n $NS rollout status deploy/${org}-${PEER_NAME}-cc-${ccname}
 }
 
 # Activate the installed chaincode but do not package/install a new archive.
 function activate_chaincode() {
   set -x
+  local ccname=$1
 
-  set_chaincode_id
-  activate_chaincode_for org1 $CHAINCODE_ID
+  set_chaincode_id ${ccname}
+  activate_chaincode_for org1 $CHAINCODE_ID $ccname
 }
 
 function activate_chaincode_for() {
   local org=$1
   local cc_id=$2
-  echo "Activating chaincode ${CHAINCODE_ID}"
+  local ccname=$3
+  echo "Activating chaincode ${cc_id}"
 
   echo 'set -x 
   export CORE_PEER_ADDRESS='${org}'-peer1:7051
@@ -120,21 +144,21 @@ function activate_chaincode_for() {
   peer lifecycle \
     chaincode approveformyorg \
     --channelID '${CHANNEL_NAME}' \
-    --name '${CHAINCODE_NAME}' \
+    --name '${ccname}' \
     --version 1 \
     --package-id '${cc_id}' \
     --sequence 1 \
     -o org0-orderer1:6050 \
-    --tls --cafile /var/hyperledger/fabric/organizations/ordererOrganizations/org0.example.com/msp/tlscacerts/org0-tls-ca.pem
+    --tls false --cafile /var/hyperledger/fabric/organizations/ordererOrganizations/org0.example.com/msp/tlscacerts/org0-tls-ca.pem
   
   peer lifecycle \
     chaincode commit \
     --channelID '${CHANNEL_NAME}' \
-    --name '${CHAINCODE_NAME}' \
+    --name '${ccname}' \
     --version 1 \
     --sequence 1 \
     -o org0-orderer1:6050 \
-    --tls --cafile /var/hyperledger/fabric/organizations/ordererOrganizations/org0.example.com/msp/tlscacerts/org0-tls-ca.pem
+    --tls false --cafile /var/hyperledger/fabric/organizations/ordererOrganizations/org0.example.com/msp/tlscacerts/org0-tls-ca.pem
   ' | exec kubectl -n $NS exec deploy/${org}-admin-cli -c main -i -- /bin/bash
 }
 
@@ -142,24 +166,34 @@ function activate_chaincode_for() {
 function deploy_chaincode() {
   set -x
 
-  install_chaincode
-  launch_chaincode_service org1 peer1
-  launch_chaincode_service org1 peer2
-  activate_chaincode
+  local ccimage=$1
+  local ccname=$2
+
+  install_chaincode ${ccname}
+  
+  launch_chaincode_service org1 peer1 ${ccimage} ${ccname}
+  #launch_chaincode_service org1 peer2 ${ccimage} ${ccname}
+  activate_chaincode ${ccname}
 }
 
 function invoke_chaincode() {
-  export parameters=$1
+  local ccname=$1
+  export parameters=$2
 
+  echo "Chaincode Name=$ccname"
   echo "Parameters=$parameters"
 
+  export CHAINCODE_NAME=$ccname
   timestamp=$(date -u +%Y%m%dT%H%M%SZ)
-  populateTemplate invoke_chaincode_template.sh ${CHANNEL_TMP_DIR}/${timestamp}_invoke_chaincode_${CHAINCODE_NAME}_${CHANNEL_NAME}_.sh
+  populateTemplate invoke_chaincode_template.sh ${CHANNEL_TMP_DIR}/${timestamp}_invoke_chaincode_${ccname}_${CHANNEL_NAME}_.sh
   cat ${populatedTemplate} | exec kubectl -n $NS exec deploy/org1-admin-cli -c main -i -- /bin/bash
 }
 
 function query_chaincode() {
-  export parameters=$1
-  populateTemplate query_chaincode_template.sh ${CHANNEL_TMP_DIR}/query_chaincode_${CHAINCODE_NAME}_${CHANNEL_NAME}.sh
+  local ccname=$1
+  export parameters=$2
+  
+  export CHAINCODE_NAME=$ccname
+  populateTemplate query_chaincode_template.sh ${CHANNEL_TMP_DIR}/query_chaincode_${ccname}_${CHANNEL_NAME}.sh
   cat ${populatedTemplate} | exec kubectl -n $NS exec deploy/org1-admin-cli -c main -i -- /bin/bash
 }
